converter/requirements.txt
python-hcl2==4.3.2
boto3>=1.34.23
botocore>=1.34.23
click>=8.1.7
